Make things package-private and move test classes to same package as source
