+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HandshakeResponseMessage.java
@@ -3,7 +3,7 @@
 
 package com.microsoft.aspnet.signalr;
 
-public class HandshakeResponseMessage {
+ class HandshakeResponseMessage {
     public String error;
 
     public HandshakeResponseMessage() {
