+++ clients/java/signalr/src/test/java/com/microsoft/aspnet/signalr/HandshakeProtocolTest.java
@@ -1,15 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-package com.microsoft.aspnet.signalr.test;
+package com.microsoft.aspnet.signalr;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
-
-import com.microsoft.aspnet.signalr.HandshakeProtocol;
-import com.microsoft.aspnet.signalr.HandshakeRequestMessage;
-import com.microsoft.aspnet.signalr.HandshakeResponseMessage;
 
 public class HandshakeProtocolTest {
