+++ clients/java/signalr/src/test/java/com/microsoft/aspnet/signalr/HubExceptionTest.java
@@ -1,13 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-package com.microsoft.aspnet.signalr.test;
+package com.microsoft.aspnet.signalr;
 
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
-
-import com.microsoft.aspnet.signalr.HubException;
 
 public class HubExceptionTest {
     @Test
