+++ clients/java/signalr/src/test/java/com/microsoft/aspnet/signalr/NegotiateResponseTest.java
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-package com.microsoft.aspnet.signalr.test;
+package com.microsoft.aspnet.signalr;
 
 import static org.junit.Assert.*;
